Split delete test suite
